refactor license struct
